chrome/gpio/requirements.txt
RPi.GPIO==0.5.11
tornado==4.2.1
